+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 13.07.2024
 * **modified:** Users can no longer rate their own setups.
 * **added:** Users can now delete their own ratings.
+* **fixed:** If a trainer setup is deleted, delete all associated ratings too.